pyproject.toml
[tool.poetry]
name = "raspberries-core"
version = "0.1.0"
description = "Backend core for raspverries"
authors = ["igor.evtushenko <igor.evtushenko@mati.io>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
asyncio = "^3.4.3"
SQLAlchemy = {extras = ["async"], version = "^1.4.39"}
pydantic = "^1.9.1"
python-environ = "^0.4.54"
bcrypt = "^3.2.2"
passlib = "^1.7.4"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
